Refactor color utilities into shared module
